test: fix e2e drop index modal selector
