--- packages/compass-indexes/src/components/drop-index-form/index.ts
@@ -1,4 +1,0 @@
-import DropIndexForm, { DropIndexFormProps } from './drop-index-form';
-
-export default DropIndexForm;
-export { DropIndexFormProps };
