--- packages/compass-indexes/src/components/drop-index-modal/index.ts
@@ -1,2 +1,0 @@
-import DropIndexModal from './drop-index-modal';
-export default DropIndexModal;
